fix: INFRA-1610 refactor helper functions, add docs
